backend/requirements.txt
flask==2.0.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0
pyjwt==2.1.0
werkzeug==2.0.1
docker-py==1.10.6
requests==2.26.0
pillow==9.5.0 
python-multipart==0.0.5